refactor: HypothesisDetails component
